% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loop_counter.R
\name{loop_counter}
\alias{loop_counter}
\title{Loop counter}
\usage{
loop_counter(i, n)
}
\arguments{
\item{i}{integer indicating the ith iteration}

\item{n}{integer indicating number of iterations in the loop}
}
\description{
Loop counter
}
\examples{
n <- 10

for (i in 1:n) {

  Sys.sleep(0.2)

  loop_counter(i = i, n = n)

}
}
