% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r_squared_for_glm.R
\name{r_squared_for_glm}
\alias{r_squared_for_glm}
\title{R squared for glm}
\usage{
r_squared_for_glm(obj)
}
\arguments{
\item{obj}{glm object}
}
\value{
R squared for glm (numeric value)
}
\description{
R squared for glm
}
\details{
R squared for glm is defined as 1 - (D_M / D_0) where D_M and D_0 is the deviance for respectively the chosen model and the null model
}
\examples{
set.seed(1)
df <- data.frame(x = 1:100, eps = rnorm(n = 100))
df$y <- floor(exp(df$x + df$eps))
m <- glm(y ~ x, data = df, family = poisson())
r_squared_for_glm(obj = m)
}
