% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove_everything.R
\name{remove_brackets}
\alias{remove_brackets}
\title{Remove brackets}
\usage{
remove_brackets(x)
}
\arguments{
\item{x}{vector of any class}
}
\description{
Remove brackets
}
\details{
This is a helper function for cut_var designed to remove brackets after calling remove_everything_*
}
