% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove_everything.R
\name{remove_everything_after}
\alias{remove_everything_after}
\title{Remove everything after desired pattern}
\usage{
remove_everything_after(pattern, x)
}
\arguments{
\item{pattern}{character string to search for}

\item{x}{vector of any class}
}
\description{
Remove everything after desired pattern
}
