% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove_everything.R
\name{remove_everything_before}
\alias{remove_everything_before}
\title{Remove everything before desired pattern}
\usage{
remove_everything_before(pattern, x)
}
\arguments{
\item{pattern}{character string to search for}

\item{x}{vector of any class}
}
\description{
Remove everything before desired pattern
}
