% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove_everything.R
\name{remove_everything_but_digits}
\alias{remove_everything_but_digits}
\title{Remove everything but digits}
\usage{
remove_everything_but_digits(x)
}
\arguments{
\item{x}{vector of any class}
}
\value{
character vector of same length as x
}
\description{
Remove everything but digits
}
